refactor(ui): move shared primitives into component layer
