+++ src/engines/ChatPanel/InputArea/components/QueuedMessages.tsx
@@ -34,6 +34,7 @@ import {
   CHAT_COMPOSER_STACK_BAR_SURFACE_BG_CLASS,
 } from "@src/config/composerStackTokens";
 import { useWebViewSensors } from "@src/lib/dndKit";
+import { messageQueueReorderActiveRef } from "@src/shared/dnd/dragSideChannel";
 import {
   type QueuedMessage,
   queueEditTargetAtom,
@@ -46,7 +47,7 @@ import QueuedMessageItem from "./QueuedMessageItem";
  * Module-level flag — set synchronously in onDragStart, cleared in onDragEnd.
  * Accessible by global drag detection to skip file drop overlay during reorder.
  */
-export const reorderActiveRef = { current: false };
+export const reorderActiveRef = messageQueueReorderActiveRef;
 
 export interface QueuedMessagesProps {
   messages: QueuedMessage[];
